refactor: 명칭 변경
